[SNAP-1608][SNAP-369](snap-engine): Fix issue with wrong or broken polygon subsets from rounding the pixel coordinates which produces NaN values on conversion from Image coordinates to Map coordinates, by converting all pixel coordinates to integer.

+++ snap-core/src/main/java/org/esa/snap/core/dataio/ProductSubsetByPolygon.java
@@ -365,7 +365,7 @@ public class ProductSubsetByPolygon {
                 throw new IllegalArgumentException("Coordinate: " + geoPos.getLonString() + "," + geoPos.getLatString() + " is not a valid Geo coordinate (Lat,Lon).");
             }
             final PixelPos pixelPos = geoCoding.getPixelPos(geoPos, null);
-            pixelCoordinates.add(new Coordinate(pixelPos.getX(), pixelPos.getY()));
+            pixelCoordinates.add(new Coordinate((int) pixelPos.getX(), (int) pixelPos.getY()));
         }
         if (!geoCoordinates[0].equals(geoCoordinates[geoCoordinates.length - 1])) {
             pixelCoordinates.add(pixelCoordinates.get(0));
